project-1/requirements.txt
langchain>=0.0.267
langchain_openai>=0.0.2
langchain_community>=0.0.8
openai>=1.3.0
chromadb>=0.4.18
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.1
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
pyyaml>=6.0 